requirements.txt
paho-mqtt
pymodbus
pyserial
